+++ lib/Command/RenderClusterVideoLandscapeCommand.php
@@ -7,6 +7,7 @@ use OCA\Journeys\Model\ClusterVideoSelection;
 use OCA\Journeys\Service\ClusterVideoFilePreparer;
 use OCA\Journeys\Service\ClusterVideoImageProvider;
 use OCA\Journeys\Service\ClusterVideoRendererLandscape;
+use OCP\IConfig;
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputArgument;
 use Symfony\Component\Console\Input\InputInterface;
@@ -20,6 +21,7 @@ class RenderClusterVideoLandscapeCommand extends Command {
         private ClusterVideoImageProvider $imageProvider,
         private ClusterVideoFilePreparer $filePreparer,
         private ClusterVideoRendererLandscape $videoRenderer,
+        private IConfig $config,
     ) {
         parent::__construct(static::$defaultName);
     }
@@ -37,6 +39,7 @@ class RenderClusterVideoLandscapeCommand extends Command {
             ->addOption('output', 'o', InputOption::VALUE_REQUIRED, 'Output mp4 path (absolute inside server)')
             ->addOption('no-audio', null, InputOption::VALUE_NONE, 'Disable soundtrack')
             ->addOption('no-motion', null, InputOption::VALUE_NONE, 'Disable GCam motion video inclusion')
+            ->addOption('no-title', null, InputOption::VALUE_NONE, 'Disable cluster name title overlay (enabled by default)')
             ->addOption('ffmpeg-verbose', null, InputOption::VALUE_NONE, 'Enable verbose FFmpeg output');
     }
 
@@ -60,6 +63,15 @@ class RenderClusterVideoLandscapeCommand extends Command {
         $disableAudio = (bool)$input->getOption('no-audio');
         $includeMotion = !(bool)$input->getOption('no-motion');
         $verbose = (bool)$input->getOption('ffmpeg-verbose');
+
+        // Determine if title should be shown: CLI flag overrides user setting
+        $showTitle = true;
+        if ($input->getOption('no-title')) {
+            $showTitle = false;
+        } else {
+            // Read user's preference (default: true)
+            $showTitle = (bool)((int)$this->config->getUserValue($user, 'journeys', 'showVideoTitle', 1));
+        }
 
         try {
             $selection = $this->imageProvider->getSelectedImagesForAlbumId($user, (int)$albumId, $minGap, $maxImages);
@@ -118,6 +130,7 @@ class RenderClusterVideoLandscapeCommand extends Command {
                 $preferredFileName,
                 $includeMotion,
                 $verbose,
+                $showTitle ? $selection->clusterName : null,
             );
             $output->writeln('<info>ffmpeg finished.</info>');
         } catch (\Throwable $e) {